Fixed icons and updated template folders order
